+++ CHANGELOG.md
@@ -55,6 +55,20 @@
 - [v0.1.0](#v010)
 
 ## Unreleased
+<<<<<<< HEAD
+=======
+
+### Fixes
+
+- On-prem gateway: generate a distinct Kong route for each match when its parent
+ `HTTPRoute` rule contains `ReplacePrefixMatch` typed `URLRewrite` filter or
+ `requestRedirect` filter.
+  ** This change will delete the combined Kong routes created for the matches with
+  these filters in their parent rules and create new distinct ones.
+ [#5521](https://github.com/Kong/kong-operator/pull/5521)
+
+## [v2.3.0]
+>>>>>>> 7eaf4131f (fix(ingress-controller): Split httproute match by path when they have PathPrefixRewrite filter or RequestRedirect filter (#5521))
 
 - HybridGateway: merge the current Route reference with existing
   `hybrid-routes` annotations during state enforcement so Routes that share a